feat: create cache table
