+++ src/big-query-support/manage-tables/index.ts
@@ -1,1 +1,3 @@
 export * from './positions-table';
+export * from './cache-table';
+export * from './get-table';
